+++ lib/main.dart
@@ -38,6 +38,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      title: "Budget App",
       debugShowCheckedModeBanner: false,
       // theme: ThemeData(useMaterial3: true),
       home: ResponsiveHandler(),
